+++ apps/community/src/apis/about/club/queries.ts
@@ -1,0 +1,14 @@
+import ClubService from './club-service';
+
+const queryKeys = {
+  all: ['clubs'] as const,
+};
+
+const clubQueryOptions = {
+  all: () => ({
+    queryKey: queryKeys.all,
+    queryFn: () => ClubService.getClubs(),
+  }),
+};
+
+export { clubQueryOptions };
