+++ apps/community/src/app/about/club/page.tsx
@@ -1,13 +1,17 @@
+import { clubQueryOptions } from '~/apis/about/club/queries';
+import { Hydrate, getDehydratedQuery } from '~/utils/react-query';
+
 import { ClubCard } from '~/components/about/club/club-card';
 import { ClubList } from '~/components/about/club/club-list';
 import { PageHeader } from '~/components/page-header';
-import { getClubs } from './remotes';
 
 // TODO: for mocking but will be replaced with a proper solution later
 export const dynamic = 'force-dynamic';
 
 export default async function ClubPage() {
-  const { data } = await getClubs();
+  const { queryKey, queryFn } = clubQueryOptions.all();
+  const query = await getDehydratedQuery({ queryKey, queryFn });
+  const { data } = query.state.data ?? {};
 
   return (
     <>
@@ -15,11 +19,13 @@ export default async function ClubPage() {
         title="동아리 소개"
         description="경기대학교 AI컴퓨터공학부의 연구실을 소개해요."
       />
-      <ClubList>
-        {data.map((club) => (
-          <ClubCard key={`club-${club.name}`} {...club} />
-        ))}
-      </ClubList>
+      <Hydrate state={{ query }}>
+        <ClubList>
+          {data?.map((club) => (
+            <ClubCard key={`club-${club.name}`} {...club} />
+          ))}
+        </ClubList>
+      </Hydrate>
     </>
   );
 }
