+++ apps/community/src/types/club.ts
@@ -1,0 +1,8 @@
+interface Club {
+  name: string;
+  description: string;
+  link?: string;
+  image?: string;
+}
+
+export type { Club };
